Patch Clone for older versions of go.
