+++ gorequest_transport_go1.3.go
@@ -1,0 +1,26 @@
+// +build go1.3, !go1.4
+
+package gorequest
+
+import (
+    "net/http"
+)
+
+// does a shallow clone of the transport
+func (s *SuperAgent) safeModifyTransport() {
+    if !s.isClone {
+        return
+    }
+    oldTransport := s.Transport
+    s.Transport = &http.Transport{
+        Proxy:                  oldTransport.Proxy,
+        Dial:                   oldTransport.Dial,
+        TLSClientConfig:        oldTransport.TLSClientConfig,
+        DisableKeepAlives:      oldTransport.DisableKeepAlives,
+        DisableCompression:     oldTransport.DisableCompression,
+        MaxIdleConnsPerHost:    oldTransport.MaxIdleConnsPerHost,
+        ResponseHeaderTimeout:  oldTransport.ResponseHeaderTimeout,
+        // new in 1.3
+        TLSHandshakeTimeout:    oldTransport.TLSHandshakeTimeout,
+    }
+}
